fix(SDK-5985): announce HTML banner overlay to TalkBack and share swipe gesture
Quick-win follow-ups to SDK-5976/SDK-5977:

- Accessibility: CTInAppHtmlBannerOverlay now sets ViewCompat.setAccessibilityPaneTitle
  on the overlay root (R.string.ct_inapp_message_shown), so the banner is announced to
  TalkBack when shown — parity with the fragment path (CTInAppBaseFragment).
- De-duplication: extract the swipe-to-dismiss gesture (fling detection, slide+fade
  animation, SWIPE_MIN_DISTANCE/SWIPE_THRESHOLD_VELOCITY) into a shared
  PartialHtmlInAppGestureListener used by both CTInAppHtmlBannerOverlay and
  CTInAppBasePartialHtmlFragment (mirrors the extraction from PR #841). The overlay
  passes onSwipeStart to set its animatingDismiss flag; the fragment triggers
  triggerSwipeDismissAction. Net -94 lines of duplicated animation code.
