+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/fragment/CTInAppBasePartialHtmlFragment.kt
@@ -5,75 +5,24 @@ import android.content.Context
 import android.content.res.Configuration
 import android.os.Bundle
 import android.view.GestureDetector
-import android.view.GestureDetector.SimpleOnGestureListener
 import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.View
 import android.view.View.OnLongClickListener
 import android.view.View.OnTouchListener
 import android.view.ViewGroup
-import android.view.animation.AlphaAnimation
-import android.view.animation.Animation
-import android.view.animation.AnimationSet
-import android.view.animation.TranslateAnimation
 import com.clevertap.android.sdk.CTWebInterface
 import com.clevertap.android.sdk.CleverTapAPI
 import com.clevertap.android.sdk.inapp.CTInAppWebView
 import com.clevertap.android.sdk.inapp.InAppWebViewClient
-import kotlin.math.abs
+import com.clevertap.android.sdk.inapp.PartialHtmlInAppGestureListener
 
 internal abstract class CTInAppBasePartialHtmlFragment : CTInAppBasePartialFragment(),
     OnTouchListener,
     OnLongClickListener {
 
-    private inner class GestureListener : SimpleOnGestureListener() {
-        override fun onFling(
-            e1: MotionEvent?,
-            e2: MotionEvent,
-            velocityX: Float,
-            velocityY: Float
-        ): Boolean {
-            if (e1 != null) {
-                if (e1.x - e2.x > SWIPE_MIN_DISTANCE && abs(velocityX.toDouble()) > SWIPE_THRESHOLD_VELOCITY) {
-                    // Right to left
-                    return remove(false)
-                } else if (e2.x - e1.x > SWIPE_MIN_DISTANCE && abs(velocityX.toDouble()) > SWIPE_THRESHOLD_VELOCITY) {
-                    // Left to right
-                    return remove(true)
-                }
-            }
-            return false
-        }
-
-        fun remove(ltr: Boolean): Boolean {
-            val animSet = AnimationSet(true)
-            val anim = if (ltr) {
-                TranslateAnimation(0f, getScaledPixels(50).toFloat(), 0f, 0f)
-            } else {
-                TranslateAnimation(0f, -getScaledPixels(50).toFloat(), 0f, 0f)
-            }
-            animSet.addAnimation(anim)
-            animSet.addAnimation(AlphaAnimation(1f, 0f))
-            animSet.setDuration(300)
-            animSet.setFillAfter(true)
-            animSet.isFillEnabled = true
-            animSet.setAnimationListener(object : Animation.AnimationListener {
-                override fun onAnimationEnd(animation: Animation?) {
-                    triggerSwipeDismissAction()
-                }
-
-                override fun onAnimationRepeat(animation: Animation?) {
-                }
-
-                override fun onAnimationStart(animation: Animation?) {
-                }
-            })
-            webView?.startAnimation(animSet)
-            return true
-        }
-    }
-
     private lateinit var gd: GestureDetector
+    private lateinit var gestureListener: PartialHtmlInAppGestureListener
 
     private var webView: CTInAppWebView? = null
 
@@ -84,7 +33,11 @@ internal abstract class CTInAppBasePartialHtmlFragment : CTInAppBasePartialFragm
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        gd = GestureDetector(context, GestureListener())
+        gestureListener = PartialHtmlInAppGestureListener(
+            scaledPixels = ::getScaledPixels,
+            onSwipeDismiss = ::triggerSwipeDismissAction
+        )
+        gd = GestureDetector(context, gestureListener)
     }
 
     override fun onCreateView(
@@ -141,6 +94,7 @@ internal abstract class CTInAppBasePartialHtmlFragment : CTInAppBasePartialFragm
                 inAppNotification.aspectRatio
             )
             this.webView = webView
+            gestureListener.webView = webView
             val webViewClient = InAppWebViewClient(this)
             webView.setWebViewClient(webViewClient)
             // Attach the swipe/pan gesture only when swipe-to-dismiss is enabled and there is no close
@@ -169,10 +123,5 @@ internal abstract class CTInAppBasePartialHtmlFragment : CTInAppBasePartialFragm
         webView.updateDimension()
         val html = inAppNotification.html ?: return
         webView.loadInAppHtml(html)
-    }
-
-    companion object {
-        private const val SWIPE_MIN_DISTANCE = 120
-        private const val SWIPE_THRESHOLD_VELOCITY = 200
     }
 }
